+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### ADDED:
+- Add sort order to banner
+
 ### CHANGED:
 - Refactored banner controller